libchromeos: allow child process option to not inherit signal mask
Currently, child processes always inherit signal mask from the parent
process. This can cause unintended side effects in the child process,
where the signals that are set in the parent's signal mask are always
blocked in the child process, so it is not able to receive those signals.

The default value for this option is set to false, which allows the child
process to receive all signals and make its own arrangement for signal
handling.

When this option is set to false, the child process will reset its own
copy of signal mask (which will unblock all signals) before executing
the program.

BUG=chromium:438778
TEST=manual test
Manual Test:
Start an AP service using apmanager which would spawn off a new process
for hostapd, then terminate the apmanager and verify the hostapd is killed
by the SIGTERM signal.

Change-Id: I4893ad6ec308375bb3a0106c2ee2c979a2fee229
Reviewed-on: https://chromium-review.googlesource.com/233185
Reviewed-by: Peter Qiu <zqiu@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>
Tested-by: Peter Qiu <zqiu@chromium.org>