libchromeos: support for DBus Property Set method
Added access mode field to each property and default it to read-only.
Also added a validator callback to each ExportedProperty. When the callback
is set, it will be used to validate the property value when setting it by
the remote application.

an upcoming CL will update the adaptor generator to generate codes to update
access mode and set the validator callback for properties that have write
access.

BUG=chromium:440139
TEST=USE="asan clang" FEATURES=test emerge-$BOARD libchromeos buffet
     apmanager peerd

Change-Id: I4d99263f8fac12bd7e64ba055731321e25184dba
Reviewed-on: https://chromium-review.googlesource.com/234051
Tested-by: Peter Qiu <zqiu@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>