privetd: Add support for connectivity monitoring
ShillClient now instantiates a ManagerProxy in order to watch its device
list.  For each device in the device list, ShillClient instantiates a
DeviceProxy and subscribes to updates about the SelectedService property
on the Device.  When the SelectedService is populated (!= "/") we
instantiate a ServiceProxy and subscribe to state updates.  We keep
a cache of the current state update.

For each proxy that we're watching property updates on, we subscribe for
change signals, and read the initial value of the property only when the
subscription finishes.

The other additional wrinkle here is that we share the ServiceProxy
between the set of device selected services and the connecting service.

BUG=chromium:441568
TEST=Integration tests now pass.

Change-Id: Ifea1674c336378df7fc6da6c78f074a42a74f5d9
Reviewed-on: https://chromium-review.googlesource.com/235650
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Vitaly Buka <vitalybuka@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>