crosh: fix noglob fix
The dash man page is not well written and makes it look like it expects:
	set -f noglob
When it really just wants the -f.  The noglob part is turned into $1.

Normally this doesn't matter as crosh doesn't process $@, but when you
run it locally, you can pass it --dev to get a dev shell.

BUG=chromium:439768
TEST=`./crosh --dev` works again
TEST=`echo *` still doesn't expand the glob

Change-Id: I7fbd8baf1f0970c1b3c8274f991ea2ed60387a63
Reviewed-on: https://chromium-review.googlesource.com/235460
Reviewed-by: Kees Cook <keescook@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>