libchromeos: Add async http I/O API, phase I.
Add asynchronous HTTP API to libchromeos and updated buffet to use
the appropriate async function for its DoCloudRequest method.

The actual implementation of asynchronous I/O for CURL transport
is deferred to a following CL to limit the amount of modifications
in one change.

BUG=chromium:427963
TEST=FEATURES=test emerge-link libchromeos peerd privetd buffet

Change-Id: If9ba6b4a27d0a99c29feb1cc6d594532f333e0b0
Reviewed-on: https://chromium-review.googlesource.com/238930
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Anton Muhin <antonm@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>