init: Remove /var/run and /var/lock symlinks on shutdown.
This fixes loss of stateful partition after auto-update rollback from
R38 or later to R37 or earlier.  Only cases are covered in which
rollback is initiated after a clean shutdown of the new OS, however
paradoxically this is the majority of cases due to
https://crbug.com/397602 in conjunction with widespread use of
charging carts that perform regular power cycles.

BUG=chromium:447227
TEST=manually tested reboot R39 --> R39
TEST=manually tested rollback R39 --> R37
TEST=manually tested that there is no crash when removing symlink
  while it holds a lazily umounted but still live fs

Change-Id: Ifc755fc00a549f17d542f5d2f9d880bcf36b25ca
Reviewed-on: https://chromium-review.googlesource.com/239870
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Thiemo Nagel <tnagel@chromium.org>
Tested-by: Thiemo Nagel <tnagel@chromium.org>