login: Move remaining session_manager setup into session_manager_main.cc
After moving all the chrome-specific flag setup stuff out of ui.conf
and into C++ classes, there were still a few session_manager-specific
flags that we still determined the values of in ui.conf. This is
fine until one needs to tell upstart to 'expect fork' in order to
track the PID of the session_manager -- which is needed in order
to sandbox the UI the way we'd like to.

This CL gets rid of those setup stanzas, which involve shelling out
to some tools and therefore confuse 'expect fork'. The code is moved
inside session_manager_main.cc and implemented in C++.

BUG=None
TEST=login_* autotests

Change-Id: I8de34bb32eb32e66e3044d371a51fcb116ea092c
Reviewed-on: https://chromium-review.googlesource.com/241012
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>