shill: DBusAdaptor: Sanitize path elements
This CL ensures that Device RPC identifiers are valid D-Bus
pathnames.  In specific, since device RPC identifiers are
based on the device name, and valid characters for device
names are a superset of those allowed for D-Bus pathnames,
we need to sanitize the device name before creating a D-Bus
pathname.

This new method is invoked by both the Device D-Bus adaptor
and the IPConfig D-Bus adaptor, which also uses the device
name as part of its path.

BUG=chromium:449650
TEST=Unit tests

Change-Id: If80f06d5cee9b218f527b0a26bf7cf400046206e
Reviewed-on: https://chromium-review.googlesource.com/241474
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>