init: killers_unittest: update pid ns detection
We want to add pretty proctitles to the sysroot wrapper so it's easier
to pick things out of the process listing.  Update the unittest check
to handle the new format.

Also fix a typo in the ps listing.

BUG=chromium:412484
TEST=`FEATURES=test emerge-link chromeos-init` passes w/CL:241049

Change-Id: Icf4603e10de8334d7ed6a5411ebd3d18890b7936
Reviewed-on: https://chromium-review.googlesource.com/241481
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Trybot-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>