common-mk: platform2_test: set $USER correctly
This isn't normally used by tests so it shouldn't cause a problem, but
it does fix a desync when USER is set to a different account than the
one being used inside the sysroot.  You can see this when entering the
sysroot manually and looking at your prompt.

BUG=None
TEST=precq still passes

Change-Id: Icb9659d049ccd84eb9bdfb179e0a2a759ac0e3ba
Reviewed-on: https://chromium-review.googlesource.com/378098
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>