common-mk: Add dependency on dbus-binding-generator binary
When updating chromeos-dbus-bindings-generator binary it is quite
tedious to verify them since the GYP settings for both proxy and
adapter generation do not include the generator binary in the
dependency list, so the header files are not re-created after
the generator has been updated.

Adding the generator binary in the input dependency of the proxy
and adapter GYP rules/actions will allow GYP/ninja to pick up
when the program is updated and re-generate the resulting dbus
header files for the project being compiled.

BUG=chromium:449176
TEST=emerge-link buffet
     (after deleting the .../var/cache/portage/chromeos-base/buffet)

Change-Id: Ida4c6ca0bc28aa09d4604f5dfa44e7a7b02f272d
Reviewed-on: https://chromium-review.googlesource.com/242040
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>