login: Tell NSS to always update the on-disk DB
It turns out that NSS has code to determine whether using a tempfile
would be an order of magnitude faster than disk and, if so, it will
do so and then flush to the backing store every 10s or so. In a real
device, this isn't going to be so. In a VM on a heavily loaded
buildbot...this might be so! We have code that depends on being
able to have an NSSDB updated in one process and read from another.
We won't for much longer (http://crbug.com/396439), but until that's
done, this will ensure our existing code path continues to function.

BUG=chromium:446177
TEST=unit tests

Change-Id: I0460c3ff8ede4a919e1d2bedb4949813e6b4d640
Reviewed-on: https://chromium-review.googlesource.com/242127
Trybot-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>