glbench: Add txt image database and presubmit check for it.
The glbench test images stored in platform2/glbench/images are listed
in two txt files installed in the DUT and used during the test. These
files need to be keep in sync the images in that directory.

This patch adds a new presubmit check to ensure these files are in
sync with the images/ directory and provides a way to fix them if
they aren't.

BUG=None
TEST="repo upload . --cbr"

Change-Id: I25f432fe11701d61e292ab18a43821baa7caa7a0
Reviewed-on: https://chromium-review.googlesource.com/242403
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>