init: chromeos_startup: create /var/log/asan/ for asan tests
Tests that crash currently attempt to write to /var/log/ directly.  While
this works if the process is running as root, all other processes merely
get denied:
  20:13:09 INFO | autoserv| AUTOTEST_STATUS::
  Actual test failure: ==28760==ERROR: Can't open file: /var/log/asan.28760

Since we cannot know the name of the log file ahead of time (so that we
might pre-create & chown & chmod it), lets create a world writable dir
instead for asan to write to.  It's also a lot simpler than trying to
update every possible test to try and seed things ahead of time.

We do this whenever the device is in dev mode so that it works when
people try to run things locally.  There shouldn't be a security concern
as we already have writable dirs in here, and the path isn't executable.

BUG=chromium:453579
TEST=precq passes

Change-Id: Id2591e2e9c362cbf2df7302be52eaa9ab1d28372
Reviewed-on: https://chromium-review.googlesource.com/244732
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>