shill: VPNService: Silence error spam
The "GetTethering" and "GetPhysicalTechnologyProperty" are D-Bus
property getters.  They are run every time a GetProperties call
is run.  As such, printing out an error message if the service
is not connected (a perfectly reasonable condition) doesn't make
sense.  Populating the error will automatically omit this property
from the dict returned to the caller.

BUG=chromium:455518
TEST=Unit tests, repro conditions of the bug above

Change-Id: I26c17da083cf916f625ae25a6d98112a512ae839
Reviewed-on: https://chromium-review.googlesource.com/246510
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>