debugd: Add object path to introspection XML
crash-reporter will be using chromeos-dbus-bindings to ingest the
debugd D-Bus introspection XML to generate proxies.  This tool
requires that the object path be set as the root node name.
The debugd daemon sets this to "/org/chromium/debugd" in the
DebugDaemon constructor, so the introspection XML should be set to
match.

BUG=brillo:88
TEST=chromeos-dbus-bindings proxy generator sets object path to \
     "/org/chromium/debugd"

Change-Id: I00331e0bf29195b41cd84d4495ab47738a5a41de
Reviewed-on: https://chromium-review.googlesource.com/246743
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Stephen Fung <stevefung@chromium.org>
Trybot-Ready: Stephen Fung <stevefung@chromium.org>
Tested-by: Stephen Fung <stevefung@chromium.org>