power: Allow backlight transitions to be preempted.
When a new backlight brightness animation is requested while
an animation is already in-progress, preempt the in-progress
animation to ensure that the new animation's duration is
honored.

This fixes an issue that can occur at shutdown where the
keyboard backlight animates to 0 over two seconds in
response to the panel backlight being turned off rather than
switching to 0 instantaneously due to the system shutting
down. The issue occurs because KeyboardBacklightController
is notified about the panel backlight change before it's
notified about the shutdown request.

BUG=chromium:456492
TEST=added tests

Change-Id: I3ab0a5d10c7c8d5beda12b2205912798300cb3f5
Reviewed-on: https://chromium-review.googlesource.com/247435
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Commit-Queue: Dan Erat <derat@chromium.org>