webserver: Add support for network interfaces in web server config
Now we have an option to make web server to respond to requests only
on specific network interfaces (e.g. LAN only, leaving WAN connection
closed).

In order to create a listening socket bound to a particular network
interface we need to use a method (setsockopt with SO_BINDTODEVICE
option) which requires root-level access rights. Because of this,
we create listening sockets for each of the protocol handler to be
bound to individual network interfaces at startup of the webserver
daemon before we drop privileges.

BUG=brillo:159
TEST=`FEATURES=test emerge-link webserver`

Change-Id: Iefd676420069d46bd7d4e54155aa26a9bd3cde67
Reviewed-on: https://chromium-review.googlesource.com/250841
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>