shill: Cleanup and fix problem with default address in vpn
When default address is set as the first value in excluded list
passed to SetParameters DBus API of ThirdPartyVpn driver, the
routing entries are not correctly setup since the route for default
address is found to be local. This CL ensures that this problem
does not happen by swapping the address when the first entry is
default. The CL further cleans up IPConfig.

BUG=chromium:459609
TEST=Unit tests and manual testing of the feature.

Change-Id: I0fa167aa6cb09e467a33f876dd98ef5a9000fbb4
Reviewed-on: https://chromium-review.googlesource.com/251990
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Prabhu Kaliamoorthi <kaliamoorthi@chromium.org>
Commit-Queue: Prabhu Kaliamoorthi <kaliamoorthi@chromium.org>