libchromeos: Add overloads for D-Bus method handlers with dbus::Message
In some cases it is useful to have additional information present in
the D-Bus message used to invoke the method handler, such as the sender's
unique name to track the lifetime of the client. Add method handler
variants that pass the original dbus::Message* to the handler as an
additional parameter.

BUG=brillo:346
TEST=`FEATURES=test emerge-link libchromeos peerd privetd buffet`

Change-Id: I0f1930bdd3393abe343fea9b54504f53084b47d5
Reviewed-on: https://chromium-review.googlesource.com/252217
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>