init: Don't sync hwclock on shutdown
tlsdated periodically syncs the network time to the HW clock for us,
so there is no need to do this explicitly.  The explicit sync predates
the introduction of tlsdated and may have been required back when NTP
was used instead of tlsdated.

This change shaves about 500ms off the shutdown and reboot times on
average, as hwclock blocked the upstart job waiting for a 1-second
boundary.

BUG=chromium:458713
TEST=none

Change-Id: I3eddbc3bf42db7b1ce23bb031335c9aa8d7ae2fb
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/254851
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>