soma: Parse container spec into a protobuf
Introduce a ContainerSpec protobuf and an RPC that will return one on
request. Change the current ContainerSpec class to ContainerSpecWrapper,
and make it a wrapper around a ContainerSpec protobuf. The API remains
the same, but it uses a proto for internal storage now.

If this is too inefficient (there may be lots of copying of values around)
then we can investigate parsing more directly into the protobuf.

BUG=brillo:304
TEST=unit

Change-Id: I6465f378a9ed35c7bdab522c3aa31e713f11ee87
Reviewed-on: https://chromium-review.googlesource.com/259830
Reviewed-by: Lee Campbell <leecam@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Trybot-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>