installer: Show "Filesystem verification hash failed" even when ignored.
Even when verification is disabled with --noenable_rootfs_verification
the rootfs hash should still be correct, and it usually is. It is OK
to ignore an incorrect rootfs hash when Verified Boot is disabled but
there is no good reason to ever *silence* such a problem. This change
makes the log consistent and predictable.

Also change the misnamed "warn" boolean to "enforce_rootfs_verif" and
fix the "digest, expected" order in the log message.

BUG=None
TEST=./build_image --noenable_rootfs_verification
./bin/cros_make_image_bootable ../build/images/${BOARD}/latest/ \
                             chromiumos${IMG_TYPE}_image.bin
<boot the image>
chromeos-install --dst /dev/sda

The second TEST step rebuilds an image many times faster than the
first one, very useful for development. Unfortunately it breaks the
rootfs_hash for some unknown reason and this patch stops hiding it.

Change-Id: Ib4f2964b95fe0dc12b87aafe4ae14a4ab95bba3d
Reviewed-on: https://chromium-review.googlesource.com/257163
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Marc Herbert <marc.herbert@gmail.com>
Tested-by: Marc Herbert <marc.herbert@gmail.com>