login: Implement new-style state key generation.
This changes the code to use the new method to generate state keys if
the stable device secret is present on the device. The code falls back
to the old method if the stable device secret is not available.

BUG=chromium:466388
TEST=unit tests
CQ-DEPEND=CL:260142

Change-Id: Ieb7321aa9b0bc8718b3ed7f0d199df04c8ee73b5
Reviewed-on: https://chromium-review.googlesource.com/259971
Trybot-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Mattias Nissler <mnissler@chromium.org>